refactor(base): move output print helpers from source_io to source_base
The `output` class (output.{h,cpp}) is a pure formatter that only operates on
source_base types (realArray, matrix, matrix3, ComplexMatrix). It was sitting
in source_io/module_output but has no dependency on anything above source_base,
so it belongs in source_base as a leaf utility.

Relocating it removes a batch of source_cell -> source_io (and source_pw/
source_psi -> source_io) reverse edges that existed only to reach this printer,
without changing any behavior or namespace.

- git mv source_io/module_output/output.{h,cpp} -> source_base/
- fix the relocated header's own includes (drop ../../source_base/ prefix)
- repoint all includers to source_base/output.h
- move the source entry from source_io to source_base CMakeLists and update
  the relative output.cpp paths in all affected unit-test CMakeLists

Verified: full abacus_basic_para executable builds and links cleanly.
