+++ source/source_cell/module_neighbor/test/CMakeLists.txt
@@ -13,7 +13,7 @@ AddTest(
   TARGET MODULE_CELL_NEIGHBOR_sltk_grid
   LIBS parameter ${math_libs} base device cell_info 
   SOURCES sltk_grid_test.cpp ../sltk_grid.cpp ../sltk_atom.cpp
-    ../../../source_io/module_output/output.cpp
+
 )
 
 AddTest(
@@ -21,5 +21,5 @@ AddTest(
   LIBS parameter ${math_libs} base device cell_info 
   SOURCES sltk_atom_arrange_test.cpp  ../sltk_atom_arrange.cpp ../sltk_grid_driver.cpp ../sltk_grid.cpp
   ../sltk_atom.cpp
-  ../../../source_io/module_output/output.cpp
+
 )
