+++ source/source_cell/module_symmetry/symm_analysis.cpp
@@ -1,6 +1,6 @@
 #include "symmetry.h"
 #include "source_io/module_parameter/parameter.h"
-#include "source_io/module_output/output.h"
+#include "source_base/output.h"
 
 using namespace ModuleSymmetry;
