+++ source/source_cell/test/CMakeLists.txt
@@ -46,40 +46,40 @@ AddTest(
   TARGET MODULE_CELL_read_pp
   LIBS parameter  ${math_libs} base device
   SOURCES read_pp_test.cpp ../atom_pseudo.cpp ../pseudo.cpp ../read_pp.cpp ../read_pp_complete.cpp ../read_pp_upf201.cpp ../read_pp_upf100.cpp ../read_pp_vwr.cpp ../read_pp_blps.cpp
-  ../../source_io/module_output/output.cpp
+
 )
 
 AddTest(
   TARGET MODULE_CELL_pseudo_nc
   LIBS parameter  ${math_libs} base device
   SOURCES pseudo_nc_test.cpp ../pseudo.cpp ../atom_pseudo.cpp ../read_pp.cpp ../read_pp_complete.cpp ../read_pp_upf201.cpp ../read_pp_upf100.cpp ../read_pp_vwr.cpp
-  	../read_pp_blps.cpp ../../source_io/module_output/output.cpp
+  	../read_pp_blps.cpp
 )
 
 AddTest(
   TARGET MODULE_CELL_atom_pseudo
   LIBS parameter  ${math_libs} base device
   SOURCES atom_pseudo_test.cpp ../atom_pseudo.cpp ../pseudo.cpp ../read_pp.cpp ../read_pp_complete.cpp ../read_pp_upf201.cpp ../read_pp_upf100.cpp
-  	../read_pp_vwr.cpp ../read_pp_blps.cpp ../../source_io/module_output/output.cpp
+  	../read_pp_vwr.cpp ../read_pp_blps.cpp
 )
 
 AddTest(
   TARGET MODULE_CELL_atom_spec
   LIBS parameter ${math_libs} base device
   SOURCES atom_spec_test.cpp ../atom_spec.cpp ../atom_pseudo.cpp ../pseudo.cpp ../read_pp.cpp ../read_pp_complete.cpp ../read_pp_upf201.cpp
-  	../read_pp_upf100.cpp ../read_pp_vwr.cpp ../read_pp_blps.cpp ../../source_io/module_output/output.cpp
+  	../read_pp_upf100.cpp ../read_pp_vwr.cpp ../read_pp_blps.cpp
 )
 
 AddTest(
   TARGET MODULE_CELL_klist_test
   LIBS parameter  ${math_libs} base device symmetry
-  SOURCES klist_test.cpp ../klist.cpp ../parallel_kpoints.cpp ../../source_io/module_output/output.cpp ../k_vector_utils.cpp
+  SOURCES klist_test.cpp ../klist.cpp ../parallel_kpoints.cpp ../k_vector_utils.cpp
 )
 
 AddTest(
   TARGET MODULE_CELL_klist_test_para1
   LIBS parameter  ${math_libs} base device symmetry
-  SOURCES klist_test_para.cpp ../klist.cpp ../parallel_kpoints.cpp ../../source_io/module_output/output.cpp ../k_vector_utils.cpp
+  SOURCES klist_test_para.cpp ../klist.cpp ../parallel_kpoints.cpp ../k_vector_utils.cpp
 )
 
 add_test(NAME MODULE_CELL_klist_test_para4
@@ -139,26 +139,26 @@ add_test(NAME MODULE_CELL_parallel_kpoints_test
 AddTest(
   TARGET MODULE_CELL_unitcell_test
   LIBS parameter  ${math_libs} base device cell_info  symmetry
-  SOURCES unitcell_test.cpp ../../source_io/module_output/output.cpp ../../source_estate/cal_ux.cpp
+  SOURCES unitcell_test.cpp ../../source_estate/cal_ux.cpp
 
 )
 
 AddTest(
   TARGET MODULE_CELL_unitcell_test_readpp
   LIBS parameter  ${math_libs} base device cell_info
-  SOURCES unitcell_test_readpp.cpp ../../source_io/module_output/output.cpp
+  SOURCES unitcell_test_readpp.cpp
 )
 
 AddTest(
   TARGET MODULE_CELL_unitcell_test_para
   LIBS parameter  ${math_libs} base device cell_info
-  SOURCES unitcell_test_para.cpp ../../source_io/module_output/output.cpp
+  SOURCES unitcell_test_para.cpp
 )
 
 AddTest(
   TARGET MODULE_CELL_unitcell_test_setupcell
   LIBS parameter  ${math_libs} base device cell_info
-  SOURCES unitcell_test_setupcell.cpp ../../source_io/module_output/output.cpp
+  SOURCES unitcell_test_setupcell.cpp
 )
 
 add_test(NAME MODULE_CELL_unitcell_test_parallel
