+++ source/source_cell/test_pw/CMakeLists.txt
@@ -13,7 +13,7 @@ AddTest(
   SOURCES unitcell_test_pw.cpp ../unitcell.cpp ../read_atoms.cpp ../read_atoms_helper.cpp ../atom_spec.cpp ../update_cell.cpp ../bcast_cell.cpp
 	../atom_pseudo.cpp ../pseudo.cpp ../read_pp.cpp ../read_pp_complete.cpp ../read_pp_upf201.cpp ../read_pp_upf100.cpp
   ../read_stru.cpp   ../read_atom_species.cpp
-	../read_pp_vwr.cpp ../read_pp_blps.cpp ../../source_io/module_output/output.cpp
+	../read_pp_vwr.cpp ../read_pp_blps.cpp
   ../../source_estate/read_pseudo.cpp ../../source_estate/cal_nelec_nband.cpp
   ../../source_estate/read_orb.cpp    ../print_cell.cpp
   ../../source_estate/cal_wfc.cpp  ../sep.cpp   ../sep_cell.cpp
