+++ source/source_estate/module_dm/test/CMakeLists.txt
@@ -16,7 +16,7 @@ AddTest(
   ${ABACUS_SOURCE_DIR}/source_lcao/module_hcontainer/hcontainer.cpp
   ${ABACUS_SOURCE_DIR}/source_lcao/module_hcontainer/atom_pair.cpp
   ${ABACUS_SOURCE_DIR}/source_basis/module_ao/parallel_orbitals.cpp
-  ${ABACUS_SOURCE_DIR}/source_io/module_output/output.cpp
+
 )
 
 AddTest(
