+++ source/source_estate/test/CMakeLists.txt
@@ -86,7 +86,7 @@ AddTest(
   TARGET MODULE_ESTATE_charge_test
   LIBS parameter  ${math_libs} planewave_serial base device cell_info 
   SOURCES charge_test.cpp ../module_charge/charge.cpp
-    ../../source_io/module_output/output.cpp
+
 )
 
 AddTest(
@@ -95,14 +95,14 @@ AddTest(
   SOURCES charge_mixing_test.cpp  
   ../module_charge/charge_mixing.cpp ../module_charge/charge_mixing_dmr.cpp ../module_charge/charge_mixing_residual.cpp
   ../module_charge/charge_mixing_preconditioner.cpp ../module_charge/charge_mixing_rho.cpp
-  ../module_charge/charge_mixing_uspp.cpp ../../source_io/module_output/output.cpp
+  ../module_charge/charge_mixing_uspp.cpp
 )
 
 AddTest(
   TARGET MODULE_ESTATE_charge_extra
   LIBS parameter ${math_libs} base device cell_info
   SOURCES charge_extra_test.cpp ../module_charge/charge_extra.cpp ../../source_io/module_output/read_cube.cpp ../../source_io/module_output/write_cube.cpp
-  ../../source_io/module_output/output.cpp ../../source_base/module_fft/fft_bundle.cpp ../../source_base/module_fft/fft_cpu.cpp
+ ../../source_base/module_fft/fft_bundle.cpp ../../source_base/module_fft/fft_cpu.cpp
 )
 
 AddTest(
