+++ source/source_io/module_json/test/CMakeLists.txt
@@ -7,5 +7,5 @@ AddTest(
   TARGET MODULE_IO_JSON_OUTPUT_TEST
   LIBS parameter ${math_libs} base device cell_info 
   SOURCES para_json_test.cpp ../general_info.cpp ../init_info.cpp ../readin_info.cpp 
-  ../para_json.cpp ../abacusjson.cpp ../../module_output/output.cpp 
+  ../para_json.cpp ../abacusjson.cpp 
 )
