+++ source/source_io/test_serial/CMakeLists.txt
@@ -47,7 +47,7 @@ AddTest(
 AddTest(
   TARGET MODULE_IO_rho_io
   LIBS parameter ${math_libs} base device cell_info 
-  SOURCES rho_io_test.cpp ../module_output/read_cube.cpp ../module_output/write_cube.cpp ../module_output/output.cpp
+  SOURCES rho_io_test.cpp ../module_output/read_cube.cpp ../module_output/write_cube.cpp
 )
 
 AddTest(
