+++ source/source_lcao/module_ri/module_exx_symmetry/test/CMakeLists.txt
@@ -6,5 +6,5 @@ AddTest(
   LIBS base ${math_libs} device symmetry neighbor parameter
   SOURCES symmetry_rotation_test.cpp ../symmetry_rotation.cpp ../symmetry_rotation_output.cpp ../irreducible_sector.cpp ../irreducible_sector_bvk.cpp
   ../../../../source_basis/module_ao/parallel_orbitals.cpp
-  ../../../../source_io/module_output/output.cpp
+
 )
