+++ source/source_psi/test/psi_initializer_unit_test.cpp
@@ -10,7 +10,7 @@
 #include "../psi_init_random.h"
 #include "source_pw/module_pwdft/vl_pw.h"
 #include "source_cell/klist.h"
-#include "source_io/module_output/output.h"
+#include "source_base/output.h"
 
 /*
 =========================
@@ -72,7 +72,6 @@ pseudopot_cell_vl::pseudopot_cell_vl() {}
 pseudopot_cell_vl::~pseudopot_cell_vl() {}
 Magnetism::Magnetism() {}
 Magnetism::~Magnetism() {}
-void output::printM3(std::ofstream &ofs, const std::string &description, const ModuleBase::Matrix3 &m) {}
 #ifdef __LCAO
 ORB_gaunt_table::ORB_gaunt_table() {}
 ORB_gaunt_table::~ORB_gaunt_table() {}
