+++ source/source_pw/module_pwdft/test/CMakeLists.txt
@@ -33,7 +33,7 @@ AddTest(
 	LIBS parameter ${math_libs} base device planewave
 	SOURCES structure_factor_test.cpp ../structure_factor.cpp ../parallel_grid.cpp
 	../../../source_cell/unitcell.cpp
-	../../../source_io/module_output/output.cpp
+
 	../../../source_cell/update_cell.cpp
 	../../../source_cell/bcast_cell.cpp
 	../../../source_cell/print_cell.cpp
